Revert "Remove duplicated slit-edge tracing functionality"
