+++ doc/api/pypeit.core.rst
@@ -34,6 +34,7 @@ Submodules
    pypeit.core.skysub
    pypeit.core.telluric
    pypeit.core.trace
+   pypeit.core.trace_slits
    pypeit.core.tracewave
    pypeit.core.wave
