+++ doc/api/pypeit.core.trace_slits.rst
@@ -1,0 +1,7 @@
+pypeit.core.trace\_slits module
+===============================
+
+.. automodule:: pypeit.core.trace_slits
+   :members:
+   :undoc-members:
+   :show-inheritance:
