+++ doc/api/pypeit.rst
@@ -40,6 +40,7 @@ Submodules
    pypeit.tiltimage
    pypeit.traceimage
    pypeit.tracepca
+   pypeit.traceslits
    pypeit.utils
    pypeit.wavecalib
    pypeit.waveimage
