+++ doc/api/pypeit.traceslits.rst
@@ -1,0 +1,7 @@
+pypeit.traceslits module
+========================
+
+.. automodule:: pypeit.traceslits
+   :members:
+   :undoc-members:
+   :show-inheritance:
