+++ pypeit/core/pca.py
@@ -4,8 +4,6 @@ Implement principle-component-analysis tools.
 .. _numpy.ndarray: https://docs.scipy.org/doc/numpy/reference/generated/numpy.ndarray.html
 .. _sklearn.decomposition.PCA: https://scikit-learn.org/stable/modules/generated/sklearn.decomposition.PCA.html
 """
-from IPython import embed
-
 import numpy as np
 
 from matplotlib import pyplot as plt
@@ -145,7 +143,7 @@ def fit_pca_coefficients(coeff, order, ivar=None, weights=None, function='legend
             PCA component coefficients. If the PCA decomposition used
             :math:`N_{\rm comp}` components for :math:`N_{\rm vec}`
             vectors, the shape of this array must be :math:`(N_{\rm
-            vec}, N_{\rm comp})`. The array can be 1D with shape
+            comp}, N_{\rm vec})`. The array can be 1D with shape
             :math:`(N_{\rm vec},)` if there was only one PCA
             component.
         order (:obj:`int`, `numpy.ndarray`_):
@@ -157,21 +155,13 @@ def fit_pca_coefficients(coeff, order, ivar=None, weights=None, function='legend
         ivar (`numpy.ndarray`_, optional):
             Inverse variance in the PCA coefficients to use during
             the fit; see the `invvar` parameter of
-            :func:`pypeit.utils.robust_polyfit_djs`. If None, fit is
-            not error weighted. If a vector with shape :math:`(N_{\rm
-            vec},)`, the same error will be assumed for all PCA
-            components (i.e., `ivar` will be expanded to match the
-            shape of `coeff`). If a 2D array, the shape must match
-            `coeff`.
+            :func:`pypeit.utils.robust_polyfit_djs`. Must be the same
+            shape as `coeff`. If None, fit is not error weighted.
         weights (`numpy.ndarray`_, optional):
             Weights to apply to the PCA coefficients during the fit;
             see the `weights` parameter of
-            :func:`pypeit.utils.robust_polyfit_djs`. If None, the
-            weights are uniform. If a vector with shape
-            :math:`(N_{\rm vec},)`, the same weights will be assumed
-            for all PCA components (i.e., `weights` will be expanded
-            to match the shape of `coeff`). If a 2D array, the shape
-            must match `coeff`.
+            :func:`pypeit.utils.robust_polyfit_djs`. Must be the same
+            shape as `coeff`. If None, weights are uniform.
         function (:obj:`str`, optional):
             Type of function used to fit the data.
         lower (:obj:`float`, optional):
@@ -218,9 +208,7 @@ def fit_pca_coefficients(coeff, order, ivar=None, weights=None, function='legend
     """
     # Check the input
     #   - Get the shape of the input data to fit
-    _coeff = np.asarray(coeff)
-    if _coeff.ndim == 1:
-        _coeff = np.expand_dims(_coeff, 1)
+    _coeff = np.atleast_2d(coeff)
     if _coeff.ndim != 2:
         raise ValueError('Array with coefficiencts cannot be more than 2D')
     nvec, npca = _coeff.shape
@@ -229,9 +217,7 @@ def fit_pca_coefficients(coeff, order, ivar=None, weights=None, function='legend
     if _ivar is not None and _ivar.shape != _coeff.shape:
         raise ValueError('Inverse variance array does not match input coefficients.')
     #   - Check the weights
-    _weights = np.ones(_coeff.shape, dtype=float) if weights is None else np.asarray(weights)
-    if _weights.ndim == 1:
-        _weights = np.tile(_weights, (_coeff.shape[1],1)).T
+    _weights = np.ones(_coeff.shape, dtype=float) if weights is None else np.atleast_2d(weights)
     if _weights.shape != _coeff.shape:
         raise ValueError('Weights array does not match input coefficients.')
     #   - Set the abscissa of the data if not provided and check its
@@ -265,6 +251,7 @@ def fit_pca_coefficients(coeff, order, ivar=None, weights=None, function='legend
 
     inmask = np.ones_like(coo, dtype=bool)
     for i in range(npca):
+        # TODO I
         coeff_used[:,i], fit_coeff[i] \
                 = utils.robust_polyfit_djs(coo, _coeff[:,i], _order[i], inmask=inmask,
                                            invvar=None if _ivar is None else _ivar[:,i],
@@ -318,10 +305,9 @@ def pca_predict(x, pca_coeff_fits, pca_components, pca_mean, mean, function='leg
         pca_mean (`numpy.ndarray`_):
             The mean offset of the PCA decomposotion for each pixel.
             Shape is :math:`(N_{\rm pix},)`.
-        mean (:obj:`float`, `numpy.ndarray`_):
-            The mean offset of each trace coordinate to use for the
-            PCA prediction. This is typically identical to `x`, and
-            its shape must match `x`.
+        mean (`numpy.ndarray`_):
+            The mean offset applied to each vector before the PCA.
+            Shape is :math:`(N_{\rm vec},)`.
     
     Returns:
         `numpy.ndarray`_: PCA constructed vectors, one per position
@@ -329,18 +315,15 @@ def pca_predict(x, pca_coeff_fits, pca_components, pca_mean, mean, function='leg
         x},N_{\rm pix})`, depending on the input shape/type of `x`.
     """
     _x = np.atleast_1d(x)
-    _mean = np.atleast_1d(mean)
     if _x.ndim != 1:
         raise ValueError('Coordinates for predicted vectors must be no more than 1D.')
-    if _mean.shape != _x.shape:
-        raise ValueError('Input mean must match the shape of the input prediction coordinates.')
     # Calculate the coefficients using the best fitting function
     npca = pca_components.shape[0]
     c = np.zeros((_x.size, npca), dtype=float)
     for i in range(npca):
         c[:,i] = utils.func_val(pca_coeff_fits[i], _x, function)
     # Calculate the predicted vectors and return them
-    vectors = np.dot(c, pca_components) + pca_mean[None,:] + _mean[:,None]
+    vectors = np.dot(c, pca_components) + pca_mean[None,:] + mean[:,None]
     return vectors if isinstance(x, np.ndarray) else vectors[0,:]
