+++ pypeit/spectrographs/gemini_gmos.py
@@ -91,8 +91,13 @@ class GeminiGMOSSpectrograph(spectrograph.Spectrograph):
         Set default parameters for Keck LRISb reductions.
         """
         par = pypeitpar.PypeItPar()
+        par['calibrations']['slits']['sigdetect'] = 20.
+        par['calibrations']['slits']['trace_npoly'] = 3
         par['calibrations']['slitedges']['edge_thresh'] = 20.
         par['calibrations']['slitedges']['fit_order'] = 3
+        # TODO: No longer a parameter
+#        par['calibrations']['slits']['fracignore'] = 0.02
+#        par['calibrations']['slits']['pcapar'] = [3,2,1,0]
 
         # 1D wavelength solution
         par['calibrations']['wavelengths']['rms_threshold'] = 0.40  # Might be grating dependent..
