+++ pypeit/spectrographs/keck_nirspec.py
@@ -75,6 +75,7 @@ class KeckNIRSPECSpectrograph(spectrograph.Spectrograph):
         # Lower the default threshold for tilts
         par['calibrations']['tilts']['tracethresh'] = 10.
         # Slits
+        par['calibrations']['slits']['sigdetect'] = 200.
         par['calibrations']['slitedges']['edge_thresh'] = 200.
         # 1D wavelength solution
         par['calibrations']['wavelengths']['lamps']  = ['OH_R24000']
