+++ pypeit/spectrographs/lbt_mods.py
@@ -167,6 +167,7 @@ class LBTMODS1RSpectrograph(LBTMODSSpectrograph):
         par['calibrations']['wavelengths']['n_final'] = 4
 
         # slit
+        par['calibrations']['slits']['sigdetect'] = 300
         par['calibrations']['slitedges']['edge_thresh'] = 300
 
         # Set wave tilts order
@@ -244,6 +245,7 @@ class LBTMODS1BSpectrograph(LBTMODSSpectrograph):
         par['calibrations']['wavelengths']['n_first'] = 1
 
         # slit
+        par['calibrations']['slits']['sigdetect'] = 300
         par['calibrations']['slitedges']['edge_thresh'] = 300
 
         # Set wave tilts order
@@ -326,6 +328,7 @@ class LBTMODS2RSpectrograph(LBTMODSSpectrograph):
 
 
         # slit
+        par['calibrations']['slits']['sigdetect'] = 300
         par['calibrations']['slitedges']['edge_thresh'] = 300
 
         # Set wave tilts order
@@ -403,6 +406,7 @@ class LBTMODS2BSpectrograph(LBTMODSSpectrograph):
         par['calibrations']['wavelengths']['n_first'] = 1
 
         # slit
+        par['calibrations']['slits']['sigdetect'] = 300
         par['calibrations']['slitedges']['edge_thresh'] = 300
 
         # Set wave tilts order
