+++ pypeit/spectrographs/mmt_binospec.py
@@ -202,6 +202,8 @@ class MMTBINOSPECSpectrograph(spectrograph.Spectrograph):
         par = pypeitpar.PypeItPar()
         par['rdx']['spectrograph'] = 'keck_deimos'
         # Set wave tilts order
+        par['calibrations']['slits']['sigdetect'] = 50.
+        par['calibrations']['slits']['trace_npoly'] = 3
         par['calibrations']['slitedges']['edge_thresh'] = 50.
         par['calibrations']['slitedges']['fit_order'] = 3
