+++ pypeit/spectrographs/vlt_fors.py
@@ -51,6 +51,9 @@ class VLTFORSSpectrograph(spectrograph.Spectrograph):
                 par['calibrations'][key]['process']['overscan'] = 'median'
 
         # Adjustments to slit and tilts for NIR
+        par['calibrations']['slits']['sigdetect'] = 50.
+        par['calibrations']['slits']['trace_npoly'] = 3
+        par['calibrations']['slits']['maxshift'] = 0.5
         par['calibrations']['slitedges']['edge_thresh'] = 50.
         par['calibrations']['slitedges']['fit_order'] = 3
         par['calibrations']['slitedges']['max_shift_adj'] = 0.5
