+++ pypeit/tests/tstutils.py
@@ -2,15 +2,16 @@
 Odds and ends in support of tests
 """
 import os
-import copy
 import pytest
+import numpy as np
+import copy
 
 from IPython import embed
 
-import numpy as np
 from astropy import time
 
 from pypeit import arcimage
+#from pypeit import traceslits
 from pypeit import edgetrace
 from pypeit import wavecalib
 from pypeit import flatfield
@@ -114,7 +115,7 @@ def dummy_fitstbl(nfile=10, spectro_name='shane_kast_blue', directory='', notype
     return fitstbl
 
 # TODO: Need to split this into functions that do and do not require
-# cooked.
+# cooked.  We should remove the get_spectrograph option.
 def load_kast_blue_masters(aimg=False, edges=False, tilts=False, wvcalib=False, pixflat=False):
     """
     Load up the set of shane_kast_blue master frames
@@ -139,11 +140,15 @@ def load_kast_blue_masters(aimg=False, edges=False, tilts=False, wvcalib=False, 
     spectrograph.naxis = (2112,350)     # Image shape with overscan
 
     master_dir = os.path.join(os.getenv('PYPEIT_DEV'), 'Cooked', 'Shane_Kast_blue')
+#    master_dir = root_path+'_'+spectrograph.spectrograph
 
     reuse_masters = True
 
     # Load up the Masters
     ret = []
+
+#    if get_spectrograph:
+#        ret.append(spectrograph)
 
     master_key = 'A_1_01'
     if aimg:
@@ -155,10 +160,14 @@ def load_kast_blue_masters(aimg=False, edges=False, tilts=False, wvcalib=False, 
     if edges:
         trace_file = '{0}.gz'.format(os.path.join(master_dir,
                                         MasterFrame.construct_file_name('Edges', master_key)))
+#        tslits_dict, mstrace = traceslits.TraceSlits.load_from_file(trace_file)
+#        ret.append(tslits_dict)
+#        ret.append(mstrace)
         ret.append(edgetrace.EdgeTraceSet.from_file(trace_file))
 
     if tilts:
         tilts_file = os.path.join(master_dir, MasterFrame.construct_file_name('Tilts', master_key))
+        #tilts_dict = wavetilts.WaveTilts.load_from_file(tilts_file)
         tilts_dict = wavetilts.WaveTilts.from_master_file(tilts_file).tilts_dict
         ret.append(tilts_dict)
 
@@ -171,6 +180,8 @@ def load_kast_blue_masters(aimg=False, edges=False, tilts=False, wvcalib=False, 
 
     # Pixelflat
     if pixflat:
+        #flatField = flatfield.FlatField(spectrograph,
+        #                                spectrograph.default_pypeit_par()['calibrations']['pixelflatframe'])
         calib_file = os.path.join(master_dir,
                                   MasterFrame.construct_file_name('Flat', master_key))
         flatField = flatfield.FlatField.from_master_file(calib_file)
@@ -178,3 +189,31 @@ def load_kast_blue_masters(aimg=False, edges=False, tilts=False, wvcalib=False, 
 
     # Return
     return ret
+
+#def instant_traceslits(mstrace_file, det=None):
+#    """
+#    Instantiate a TraceSlits object from the master file
+#
+#    The loaded tslits_dict is set as the atribute
+#
+#    Args:
+#        mstrace_file (str):
+#        det (int, optional):
+#
+#    Returns:
+#        Spectrograph, TraceSlits:
+#
+#    """
+#    # Load
+#    tslits_dict, mstrace = traceslits.TraceSlits.load_from_file(mstrace_file)
+#    # Instantiate
+#    spectrograph = load_spectrograph(tslits_dict['spectrograph'])
+#    par = spectrograph.default_pypeit_par()
+#    msbpm = spectrograph.bpm(None, det, shape=mstrace.shape)
+#    #binning = tslits_dict['binspectral'], tslits_dict['binspatial']
+#    traceSlits = traceslits.TraceSlits(spectrograph, par['calibrations']['slits'],
+#                                       msbpm=msbpm)
+#    traceSlits.mstrace = copy.deepcopy(mstrace)
+#    traceSlits.tslits_dict = copy.deepcopy(tslits_dict)
+#    return spectrograph, traceSlits
+
