+++ requirements_doc.txt
@@ -1,4 +1,3 @@
 -r pypeit/requirements.txt
 Sphinx>=1.6.5
-sphinx-automodapi>=0.12
